chore: update Go versions (#133)
Go 1.22 has just been released, meaning that only Go 1.22 and 1.21 are
officially in support. Testing has been improved to test against these
versions using the `stable` and `oldstable` references. These will
change when new versions of Go are released, meaning that we don't need
to keep updating the GitHub Actions file.

The `go.mod` file has been updated to Go 1.20. There's value in doing
this because the dependency tracking syntax has changed since Go 1.16.
And we have been testing against Go 1.20, so we know this works. It was
decided not to go for an intermediate version (say Go 1.18) since we
haven't been testing against that, so there's potentially additional
work needed.